moves search functions in dedicated files
